requirements.txt
confluent-kafka
faust-streaming[rocksdict]
python-dotenv
